// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ROLE = {
  Admin: 'Admin',
  Employee: 'Employee',
  Manager: 'Manager'
} as const

export type ROLE = (typeof ROLE)[keyof typeof ROLE]


export const STATUS = {
  APPLICATION_RECEIVED: 'APPLICATION_RECEIVED',
  INTERVIEW_SCHEDULED: 'INTERVIEW_SCHEDULED',
  HIRED: 'HIRED',
  NOT_ACCEPTED: 'NOT_ACCEPTED'
} as const

export type STATUS = (typeof STATUS)[keyof typeof STATUS]